// limit-guard-server/src/module/playGround/playGround.interface.ts
export interface ProductInterface {
  id: number;
  name: string;
  price: number;
}

export interface StudentInterface {
  id: number;
  name: string;
  course: string;
}

export interface EmployeeInterface {
  id: number;
  name: string;
  department: string;
}

export interface MovieInterface {
  id: number;
  title: string;
  rating: number;
}

export interface BookInterface {
  id: number;
  title: string;
  author: string;
}